+++ src/Gateways/BancontactGateway.php
@@ -10,8 +10,8 @@
 
 namespace Pronamic\WordPress\Pay\Extensions\MemberPress\Gateways;
 
-use Pronamic\WordPress\Pay\Plugin;
 use Pronamic\WordPress\Pay\Core\PaymentMethods;
+use Pronamic\WordPress\Pay\Extensions\MemberPress\Pronamic;
 
 /**
  * WordPress pay MemberPress Bancontact/Mister Cash gateway
@@ -44,6 +44,6 @@ class BancontactGateway extends Gateway {
 	 * @return string
 	 */
 	protected function get_icon() {
-		return plugins_url( 'images/bancontact/icon-32x32.png', Plugin::$file );
+		return Pronamic::get_method_icon_url( $this->payment_method );
 	}
 }